Changing structure of the project
