+++ control/templates/control/index.html
@@ -1,0 +1,2 @@
+<!DOCTYPE html>
+{% extends 'control/base.html' %}
