+++ control/urls.py
@@ -15,13 +15,17 @@ Including another URLconf
 """
 from django.conf import settings
 from django.conf.urls import url, include
-
 from django.conf.urls.static import static
 from django.contrib import admin
 from django.contrib.staticfiles.views import serve
 
+from control.views import index_view
+
 urlpatterns = [
+    url(r'^$', index_view, name='index_view'),
+
     url(r'^admin/', admin.site.urls),
+
     url(r'control/', include('control.users.urls')),
 ]
